General Enhancements 17042018
Cleaning Drag and Drop leftover code, preventing self-referencing in equations, renaming equations and fixing titles.
