+++ SpaceY/ClientApp/routes.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { Redirect, Route, Switch, withRouter } from "react-router-dom";
 import { AllEquations } from "./components/AllEquations";
-import { Dndbox } from "./components/Dndbox";
 import { EquationCreator } from "./components/EquationCreator";
 import { EquationDetail } from "./components/EquationDetail";
 import { Home } from "./components/Home";
@@ -17,8 +16,7 @@ export const routes = (
         <Switch>
             <Route path="/home/:id([0-9]*)" component={withRouter(Home)}/>
             <Route path="/equationcreator/:id([0-9]*)" component={withRouter(EquationCreator)}/>
-            <Route path="/dndbox/:id([0-9]*)" component={Dndbox}/>
-            <Route path="/AllEquations/:id([0-9]*)" component={withRouter(AllEquations)}/>
+            <Route path="/allequations/:id([0-9]*)" component={withRouter(AllEquations)}/>
             <Route path="/equations/:id([0-9]+)" component={withRouter(EquationDetail)}/>
             <Redirect to="/home/"/>
         </Switch>
